+++ core/theme/src/main/java/com/dede/android_eggs/views/settings/compose/prefs/ColorSourcePrefUtil.kt
@@ -5,6 +5,7 @@ import androidx.annotation.ChecksSdkIntAtLeast
 import com.dede.android_eggs.views.settings.compose.basic.mutablePrefColorState
 import com.dede.android_eggs.views.settings.compose.basic.mutablePrefIntState
 import com.dede.android_eggs.views.theme.defaultSeedColor
+import com.dede.android_eggs.views.theme.isWallpaperColorSupported
 
 object ColorSourcePrefUtil {
 
@@ -15,10 +16,12 @@ object ColorSourcePrefUtil {
     const val KEY_COLOR_SOURCE = "pref_key_color_source_1"
     const val KEY_SEED_COLOR = "pref_key_seed_color"
 
-    val DEFAULT_SOURCE = if (isDynamicColorSupported()) SOURCE_DYNAMIC else SOURCE_DEFAULT
+    val DEFAULT_SOURCE = if (isDynamicColorSourceSupported()) SOURCE_DYNAMIC else SOURCE_DEFAULT
 
     val colorSourceState = mutablePrefIntState(KEY_COLOR_SOURCE, DEFAULT_SOURCE)
     val seedColorState = mutablePrefColorState(KEY_SEED_COLOR, defaultSeedColor)
+
+    fun isDynamicColorSourceSupported(): Boolean = isDynamicColorSupported() || isWallpaperColorSupported()
 
     @ChecksSdkIntAtLeast(api = Build.VERSION_CODES.S)
     fun isDynamicColorSupported(): Boolean {